Improve unittests (closes #10)
